Create Why-I-Chose-Blast.md
Here is the story of how I found Blast.  I think it is important to include a market analsys with every github repository.  It is the first thing that eery new user needs to read.  
I was not quite sure where to put it.  You may want to change the file name, and if you like it, link to it from the README. 

Hope that helps.  And thank you for all of he hard work in building Blast.